examples/quick-start-python/requirements.txt
flask
opentelemetry-distro
opentelemetry-exporter-otlp